added types to function headers

+++ include.jl
@@ -1,3 +1,3 @@
-function almost_equal(a, b, ϵ)
+function almost_equal(a::Float64, b::Float64, ϵ::Float64)
   return abs(a-b) <= ϵ
 end
